move default price to a conern
